display user informations on frontend
