+++ frontend/src/app/[locale]/account/page.tsx
@@ -1,0 +1,5 @@
+export default function AccountPage() {
+  return(
+    <h1>Account</h1>
+  )
+}
