--- frontend/src/app/[locale]/dashboard/page.tsx
@@ -1,5 +1,0 @@
-export default function DashBoardPage() {
-  return(
-    <h1>Dashboard</h1>
-  )
-}
